backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-docx==1.2.0
lxml==6.0.2
pydantic==2.10.6
python-multipart==0.0.20
pyyaml==6.0.2
openai==1.59.7
anthropic==0.45.1
python-dotenv==1.0.1
